refactor: consolidate thread runtime helpers
