+++ tests/Unit/backend/web/services/test_event_buffer_owner.py
@@ -3,10 +3,24 @@ from __future__ import annotations
 import importlib
 
 
-def test_event_buffer_owner_lives_under_backend_thread_runtime() -> None:
-    owner_module = importlib.import_module("backend.thread_runtime.event_buffer")
-    shell_module = importlib.import_module("backend.web.services.event_buffer")
+def test_thread_runtime_namespace_exports_legacy_helpers() -> None:
+    history_owner = importlib.import_module("backend.thread_runtime.history")
+    history_shell = importlib.import_module("backend.thread_history")
+    projection_owner = importlib.import_module("backend.thread_runtime.projection")
+    projection_shell = importlib.import_module("backend.thread_projection")
+    convergence_owner = importlib.import_module("backend.thread_runtime.convergence")
+    convergence_shell = importlib.import_module("backend.thread_runtime_convergence")
+    sandbox_owner = importlib.import_module("backend.thread_runtime.sandbox")
+    sandbox_shell = importlib.import_module("backend.thread_sandbox")
+    reads_owner = importlib.import_module("backend.thread_runtime.events.reads")
+    reads_shell = importlib.import_module("backend.run_event_reads")
+    buffer_owner = importlib.import_module("backend.thread_runtime.events.buffer")
+    buffer_shell = importlib.import_module("backend.web.services.event_buffer")
 
-    assert owner_module.ThreadEventBuffer is shell_module.ThreadEventBuffer
-    assert owner_module.RunEventBuffer is shell_module.RunEventBuffer
-    assert owner_module.__name__ == "backend.thread_runtime.event_buffer"
+    assert history_owner.build_thread_history_transport is history_shell.build_thread_history_transport
+    assert projection_owner.canonical_owner_threads is projection_shell.canonical_owner_threads
+    assert convergence_owner.inspect_owner_thread_runtime is convergence_shell.inspect_owner_thread_runtime
+    assert sandbox_owner.resolve_thread_sandbox is sandbox_shell.resolve_thread_sandbox
+    assert reads_owner.build_run_event_read_transport is reads_shell.build_run_event_read_transport
+    assert buffer_owner.ThreadEventBuffer is buffer_shell.ThreadEventBuffer
+    assert buffer_owner.RunEventBuffer is buffer_shell.RunEventBuffer
